requirements.txt
pyStixel-lib
torch
torchvision
pandas
coopscenes
einops
